moved some helper functions into utils from index.ts
