+++ src/@commonsku/utils/index.ts
@@ -1,16 +1,8 @@
 import ssr from './ssr';
+
 export * from './sizes';
+export * from './utils';
 export { default as generateColor } from './generateColor';
-
-export function isClientSide() {
-  return Boolean(typeof window !== 'undefined' && window.document);
-}
-
-export const valIsValid = (val: any) => val !== null && val !== undefined && val !== ''; // not checking for 0
-
-export const aeval = (obj:{[key: string]: any}, key: string, def: any) => {
-    return obj && obj[key] ? obj[key] : def;
-}
 
 export const truncate = (value: string, length: number) => {
     if(typeof value !== 'string') {
