+++ src/@commonsku/utils/utils.ts
@@ -1,0 +1,9 @@
+export const valIsValid = (val: any) => val !== null && val !== undefined && val !== ''; // not checking for 0
+
+export const aeval = (obj:{[key: string]: any}, key: string, def: any) => {
+    return obj && obj[key] ? obj[key] : def;
+}
+
+export function isClientSide() {
+  return Boolean(typeof window !== 'undefined' && window.document);
+}
